@@ -21,15 +21,29 @@
 		}
 	],
 	"require": {
-		"php":                ">=5.3.2",
-		"ext-curl":           "*",
-		"ext-xml":            "*",
-		"kriswallsmith/buzz": ">=0.7"
+		"php": "^5.6 || ^7.0",
+		"php-http/client-common": "^1.5",
+		"php-http/client-implementation": "^1.0",
+		"php-http/discovery": "^1.2",
+		"php-http/httplug": "^1.1",
+		"php-http/multipart-stream-builder": "^1.0",
+		"symfony/options-resolver": "^2.6 || ^3.0"
 	},
 	"require-dev": {
+		"guzzlehttp/psr7": "^1.2",
+		"php-http/guzzle6-adapter": "^1.0",
+		"php-http/mock-client": "^1.0",
 		"phpunit/phpunit": "~4.5"
 	},
 	"autoload": {
-		"psr-0": { "Gitlab\\": "lib/" }
+		"psr-4": { "Gitlab\\": "lib/Gitlab/" }
+	},
+	"autoload-dev": {
+		"psr-4": { "Gitlab\\Tests\\": "test/Gitlab/Tests/" }
+	},
+	"extra": {
+		"branch-alias": {
+			"dev-master": "9.0.x-dev"
+		}
 	}
 }
