@@ -4,32 +4,60 @@ class Issues extends AbstractApi
 {
     /**
      * @param int $project_id
-     * @param int $page
-     * @param int $per_page
-     * @param array $params
+     * @param array $parameters (
+     *
+     *     @var string $state     Return all issues or just those that are opened or closed.
+     *     @var string $labels    Comma-separated list of label names, issues must have all labels to be returned.
+     *                            No+Label lists all issues with no labels.
+     *     @var string $milestone The milestone title.
+     *     @var string scope      Return issues for the given scope: created-by-me, assigned-to-me or all. Defaults to created-by-me
+     *     @var int[]  $iids      Return only the issues having the given iid.
+     *     @var string $order_by  Return requests ordered by created_at or updated_at fields. Default is created_at.
+     *     @var string $sort      Return requests sorted in asc or desc order. Default is desc.
+     *     @var string $search    Search issues against their title and description.
+     * )
+     *
      * @return mixed
      */
-    public function all($project_id = null, $page = 1, $per_page = self::PER_PAGE, array $params = array())
+    public function all($project_id = null, array $parameters = [])
     {
+        $resolver = $this->createOptionsResolver();
+
+        $resolver->setDefined('state')
+            ->setAllowedValues('state', ['opened', 'closed'])
+        ;
+        $resolver->setDefined('labels');
+        $resolver->setDefined('milestone');
+        $resolver->setDefined('iids')
+            ->setAllowedTypes('iids', 'array')
+            ->setAllowedValues('iids', function (array $value) {
+                return count($value) == count(array_filter($value, 'is_int'));
+            })
+        ;
+        $resolver->setDefined('scope')
+            ->setAllowedValues('scope', ['created-by-me', 'assigned-to-me', 'all'])
+        ;
+        $resolver->setDefined('order_by')
+            ->setAllowedValues('order_by', ['created_at', 'updated_at'])
+        ;
+        $resolver->setDefined('sort')
+            ->setAllowedValues('sort', ['asc', 'desc'])
+        ;
+        $resolver->setDefined('search');
+
         $path = $project_id === null ? 'issues' : $this->getProjectPath($project_id, 'issues');
 
-        $params = array_intersect_key($params, array('labels' => '', 'state' => '', 'sort' => '', 'order_by' => '', 'milestone' => ''));
-        $params = array_merge(array(
-            'page' => $page,
-            'per_page' => $per_page
-        ), $params);
-
-        return $this->get($path, $params);
+        return $this->get($path, $resolver->resolve($parameters));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @return mixed
      */
-    public function show($project_id, $issue_id)
+    public function show($project_id, $issue_iid)
     {
-        return $this->get($this->getProjectPath($project_id, 'issues?iid='.$this->encodePath($issue_id)));
+        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)));
     }
 
     /**
@@ -44,53 +72,53 @@ class Issues extends AbstractApi
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param array $params
      * @return mixed
      */
-    public function update($project_id, $issue_id, array $params)
+    public function update($project_id, $issue_iid, array $params)
     {
-        return $this->put($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id)), $params);
+        return $this->put($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)), $params);
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @return mixed
      */
-    public function remove($project_id, $issue_id)
+    public function remove($project_id, $issue_iid)
     {
-        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id)));
+        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @return mixed
      */
-    public function showComments($project_id, $issue_id)
+    public function showComments($project_id, $issue_iid)
     {
-        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id)).'/notes');
+        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)).'/notes');
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param int $note_id
      * @return mixed
      */
-    public function showComment($project_id, $issue_id, $note_id)
+    public function showComment($project_id, $issue_iid, $note_id)
     {
-        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id)).'/notes/'.$this->encodePath($note_id));
+        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid)).'/notes/'.$this->encodePath($note_id));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param string|array $body
      * @return mixed
      */
-    public function addComment($project_id, $issue_id, $body)
+    public function addComment($project_id, $issue_iid, $body)
     {
         // backwards compatibility
         if (is_array($body)) {
@@ -99,70 +127,70 @@ class Issues extends AbstractApi
             $params = array('body' => $body);
         }
 
-        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/notes'), $params);
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/notes'), $params);
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param int $note_id
      * @param string $body
      * @return mixed
      */
-    public function updateComment($project_id, $issue_id, $note_id, $body)
+    public function updateComment($project_id, $issue_iid, $note_id, $body)
     {
-        return $this->put($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/notes/'.$this->encodePath($note_id)), array(
+        return $this->put($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/notes/'.$this->encodePath($note_id)), array(
             'body' => $body
         ));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param int $note_id
      * @return mixed
      */
-    public function removeComment($project_id, $issue_id, $note_id)
+    public function removeComment($project_id, $issue_iid, $note_id)
     {
-        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/notes/'.$this->encodePath($note_id)));
+        return $this->delete($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/notes/'.$this->encodePath($note_id)));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param string $duration
      */
-    public function setTimeEstimate($project_id, $issue_id, $duration)
+    public function setTimeEstimate($project_id, $issue_iid, $duration)
     {
-        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/time_estimate'), array('duration' => $duration));
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/time_estimate'), array('duration' => $duration));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      */
-    public function resetTimeEstimate($project_id, $issue_id)
+    public function resetTimeEstimate($project_id, $issue_iid)
     {
-        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/reset_time_estimate'));
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/reset_time_estimate'));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      * @param string $duration
      */
-    public function addSpentTime($project_id, $issue_id, $duration)
+    public function addSpentTime($project_id, $issue_iid, $duration)
     {
-        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/add_spent_time'), array('duration' => $duration));
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/add_spent_time'), array('duration' => $duration));
     }
 
     /**
      * @param int $project_id
-     * @param int $issue_id
+     * @param int $issue_iid
      */
-    public function resetSpentTime($project_id, $issue_id)
+    public function resetSpentTime($project_id, $issue_iid)
     {
-        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_id).'/reset_spent_time'));
+        return $this->post($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/reset_spent_time'));
     }
 
     /**
@@ -173,5 +201,16 @@ class Issues extends AbstractApi
     public function getTimeStats($project_id, $issue_iid)
     {
         return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid) .'/time_stats'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $issue_iid
+     *
+     * @return mixed
+     */
+    public function awardEmoji($project_id, $issue_iid)
+    {
+        return $this->get($this->getProjectPath($project_id, 'issues/'.$this->encodePath($issue_iid).'/award_emoji'));
     }
 }
