@@ -1,6 +1,14 @@
 <?php namespace Gitlab\Api;
 
 use Gitlab\Client;
+use Gitlab\HttpClient\Message\QueryStringBuilder;
+use Gitlab\HttpClient\Message\ResponseMediator;
+use Gitlab\Tests\HttpClient\Message\QueryStringBuilderTest;
+use Http\Discovery\StreamFactoryDiscovery;
+use Http\Message\MultipartStream\MultipartStreamBuilder;
+use Http\Message\StreamFactory;
+use Psr\Http\Message\ResponseInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Abstract class for Api classes
@@ -12,11 +20,6 @@ use Gitlab\Client;
 abstract class AbstractApi implements ApiInterface
 {
     /**
-     * Default entries per page
-     */
-    const PER_PAGE = 20;
-
-    /**
      * The client
      *
      * @var Client
@@ -24,11 +27,18 @@ abstract class AbstractApi implements ApiInterface
     protected $client;
 
     /**
-     * @param Client $client
+     * @var StreamFactory
      */
-    public function __construct(Client $client)
+    private $streamFactory;
+
+    /**
+     * @param Client $client
+     * @param StreamFactory|null $streamFactory
+     */
+    public function __construct(Client $client, StreamFactory $streamFactory = null)
     {
         $this->client = $client;
+        $this->streamFactory = $streamFactory ?: StreamFactoryDiscovery::find();
     }
 
     /**
@@ -41,6 +51,21 @@ abstract class AbstractApi implements ApiInterface
     }
 
     /**
+     * Performs a GET query and returns the response as a PSR-7 response object.
+     *
+     * @param string $path
+     * @param array $parameters
+     * @param array $requestHeaders
+     * @return ResponseInterface
+     */
+    protected function getAsResponse($path, array $parameters = array(), $requestHeaders = array())
+    {
+        $path = $this->preparePath($path, $parameters);
+
+        return $this->client->getHttpClient()->get($path, $requestHeaders);
+    }
+
+    /**
      * @param string $path
      * @param array $parameters
      * @param array $requestHeaders
@@ -48,9 +73,7 @@ abstract class AbstractApi implements ApiInterface
      */
     protected function get($path, array $parameters = array(), $requestHeaders = array())
     {
-        $response = $this->client->getHttpClient()->get($path, $parameters, $requestHeaders);
-
-        return $response->getContent();
+        return ResponseMediator::getContent($this->getAsResponse($path, $parameters, $requestHeaders));
     }
 
     /**
@@ -62,22 +85,35 @@ abstract class AbstractApi implements ApiInterface
      */
     protected function post($path, array $parameters = array(), $requestHeaders = array(), array $files = array())
     {
-        $response = $this->client->getHttpClient()->post($path, $parameters, $requestHeaders, $files);
+        $path = $this->preparePath($path);
 
-        return $response->getContent();
-    }
+        $body = null;
+        if (empty($files) && !empty($parameters)) {
+            $body = $this->streamFactory->createStream(QueryStringBuilder::build($parameters));
+            $requestHeaders['Content-Type'] = 'application/x-www-form-urlencoded';
+        } elseif (!empty($files)) {
+            $builder = new MultipartStreamBuilder($this->streamFactory);
 
-    /**
-     * @param string $path
-     * @param array $parameters
-     * @param array $requestHeaders
-     * @return mixed
-     */
-    protected function patch($path, array $parameters = array(), $requestHeaders = array())
-    {
-        $response = $this->client->getHttpClient()->patch($path, $parameters, $requestHeaders);
+            foreach ($parameters as $name => $value) {
+                $builder->addResource($name, $value);
+            }
 
-        return $response->getContent();
+            foreach ($files as $name => $file) {
+                $builder->addResource($name, fopen($file, 'r'), [
+                    'headers' => [
+                        'Content-Type' => $this->guessContentType($file),
+                    ],
+                    'filename' => basename($file),
+                ]);
+            }
+
+            $body = $builder->build();
+            $requestHeaders['Content-Type'] = 'multipart/form-data; boundary='.$builder->getBoundary();
+        }
+
+        $response = $this->client->getHttpClient()->post($path, $requestHeaders, $body);
+
+        return ResponseMediator::getContent($response);
     }
 
     /**
@@ -88,9 +124,17 @@ abstract class AbstractApi implements ApiInterface
      */
     protected function put($path, array $parameters = array(), $requestHeaders = array())
     {
-        $response = $this->client->getHttpClient()->put($path, $parameters, $requestHeaders);
+        $path = $this->preparePath($path);
 
-        return $response->getContent();
+        $body = null;
+        if (!empty($parameters)) {
+            $body = $this->streamFactory->createStream(http_build_query($parameters));
+            $requestHeaders['Content-Type'] = 'application/x-www-form-urlencoded';
+        }
+
+        $response = $this->client->getHttpClient()->put($path, $requestHeaders, $body);
+
+        return ResponseMediator::getContent($response);
     }
 
     /**
@@ -101,9 +145,11 @@ abstract class AbstractApi implements ApiInterface
      */
     protected function delete($path, array $parameters = array(), $requestHeaders = array())
     {
-        $response = $this->client->getHttpClient()->delete($path, $parameters, $requestHeaders);
+        $path = $this->preparePath($path, $parameters);
 
-        return $response->getContent();
+        $response = $this->client->getHttpClient()->delete($path, $requestHeaders);
+
+        return ResponseMediator::getContent($response);
     }
 
     /**
@@ -125,5 +171,53 @@ abstract class AbstractApi implements ApiInterface
         $path = rawurlencode($path);
 
         return str_replace('.', '%2E', $path);
+    }
+
+    /**
+     * Create a new OptionsResolver with page and per_page options.
+     *
+     * @return OptionsResolver
+     */
+    protected function createOptionsResolver()
+    {
+        $resolver = new OptionsResolver();
+        $resolver->setDefined('page')
+            ->setAllowedTypes('page', 'int')
+            ->setAllowedValues('page', function ($value) {
+                return $value > 0;
+            })
+        ;
+        $resolver->setDefined('per_page')
+            ->setAllowedTypes('per_page', 'int')
+            ->setAllowedValues('per_page', function ($value) {
+                return $value > 0 && $value <= 100;
+            })
+        ;
+
+        return $resolver;
+    }
+
+    private function preparePath($path, array $parameters = [])
+    {
+        if (count($parameters) > 0) {
+            $path .= '?'.QueryStringBuilder::build($parameters);
+        }
+
+        return $path;
+    }
+
+    /**
+     * @param $file
+     *
+     * @return string
+     */
+    private function guessContentType($file)
+    {
+        if (!class_exists(\finfo::class, false)) {
+            return 'application/octet-stream';
+        }
+        $finfo = new \finfo(FILEINFO_MIME_TYPE);
+
+        return $finfo->file($file);
     }
 }
