@@ -3,30 +3,52 @@
 class Groups extends AbstractApi
 {
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters (
+     *
+     *     @var int[]  $skip_groups   Skip the group IDs passes.
+     *     @var bool   $all_available Show all the groups you have access to.
+     *     @var string $search        Return list of authorized groups matching the search criteria.
+     *     @var string $order_by      Order groups by name or path. Default is name.
+     *     @var string $sort          Order groups in asc or desc order. Default is asc.
+     *     @var bool   $statistics    Include group statistics (admins only).
+     *     @var bool   $owned         Limit by groups owned by the current user.
+     * )
      * @return mixed
      */
-    public function all($page = 1, $per_page = self::PER_PAGE)
+    public function all(array $parameters = [])
     {
-        return $this->get('groups', array(
-            'page' => $page,
-            'per_page' => $per_page
-        ));
-    }
+        $resolver = $this->createOptionsResolver();
+        $booleanNormalizer = function ($value) {
+            return $value ? 'true' : 'false';
+        };
 
-    /**
-     * @param string $query
-     * @param int $page
-     * @param int $per_page
-     * @return mixed
-     */
-    public function search($query, $page = 1, $per_page = self::PER_PAGE)
-    {
-        return $this->get('groups?search='.$this->encodePath($query), array(
-            'page' => $page,
-            'per_page' => $per_page
-        ));
+        $resolver->setDefined('skip_groups')
+            ->setAllowedTypes('skip_groups', 'array')
+            ->setAllowedValues('skip_groups', function (array $value) {
+                return count($value) == count(array_filter($value, 'is_int'));
+            })
+        ;
+        $resolver->setDefined('all_available')
+            ->setAllowedTypes('all_available', 'bool')
+            ->setNormalizer('all_available', $booleanNormalizer)
+        ;
+        $resolver->setDefined('search');
+        $resolver->setDefined('order_by')
+            ->setAllowedValues('order_by', ['name', 'path'])
+        ;
+        $resolver->setDefined('sort')
+            ->setAllowedValues('sort', ['asc', 'desc'])
+        ;
+        $resolver->setDefined('statistics')
+            ->setAllowedTypes('statistics', 'bool')
+            ->setNormalizer('statistics', $booleanNormalizer)
+        ;
+        $resolver->setDefined('owned')
+            ->setAllowedTypes('owned', 'bool')
+            ->setNormalizer('owned', $booleanNormalizer)
+        ;
+
+        return $this->get('groups', $resolver->resolve($parameters));
     }
 
     /**
@@ -42,15 +64,16 @@ class Groups extends AbstractApi
      * @param string $name
      * @param string $path
      * @param string $description
+     * @param string $visibility
      * @return mixed
      */
-    public function create($name, $path, $description = null, $visibility_level = 0)
+    public function create($name, $path, $description = null, $visibility = 'private')
     {
         return $this->post('groups', array(
             'name' => $name,
             'path' => $path,
             'description' => $description,
-            'visibility_level' => $visibility_level
+            'visibility' => $visibility,
         ));
     }
 
@@ -84,17 +107,20 @@ class Groups extends AbstractApi
     }
 
     /**
-     * @param int $id
-     * @param int $page
-     * @param int $per_page
+     * @param int   $id
+     * @param array $parameters (
+     *
+     *     @var string $query A query string to search for members.
+     * )
+     *
      * @return mixed
      */
-    public function members($id, $page = 1, $per_page = self::PER_PAGE)
+    public function members($id, array $parameters = [])
     {
-        return $this->get('groups/'.$this->encodePath($id).'/members', array(
-            'page' => $page,
-            'per_page' => $per_page
-        ));
+        $resolver = $this->createOptionsResolver();
+        $resolver->setDefined('query');
+
+        return $this->get('groups/'.$this->encodePath($id).'/members', $resolver->resolve($parameters));
     }
 
     /**
@@ -132,5 +158,59 @@ class Groups extends AbstractApi
     public function removeMember($group_id, $user_id)
     {
         return $this->delete('groups/'.$this->encodePath($group_id).'/members/'.$this->encodePath($user_id));
+    }
+
+    /**
+     * @param $id
+     * @param array $parameters (
+     *
+     *     @var bool   $archived   Limit by archived status.
+     *     @var string $visibility Limit by visibility public, internal, or private.
+     *     @var string $order_by   Return projects ordered by id, name, path, created_at, updated_at, or last_activity_at fields.
+     *                             Default is created_at.
+     *     @var string $sort       Return projects sorted in asc or desc order. Default is desc.
+     *     @var string $search     Return list of authorized projects matching the search criteria.
+     *     @var bool   $simple     Return only the ID, URL, name, and path of each project.
+     *     @var bool   $owned      Limit by projects owned by the current user.
+     *     @var bool   $starred    Limit by projects starred by the current user.
+     * )
+     *
+     * @return mixed
+     */
+    public function projects($id, array $parameters = [])
+    {
+        $resolver = $this->createOptionsResolver();
+        $booleanNormalizer = function ($value) {
+            return $value ? 'true' : 'false';
+        };
+
+        $resolver->setDefined('archived')
+            ->setAllowedTypes('archived', 'bool')
+            ->setNormalizer('archived', $booleanNormalizer)
+        ;
+        $resolver->setDefined('visibility')
+            ->setAllowedValues('visibility', ['public', 'internal', 'private'])
+        ;
+        $resolver->setDefined('order_by')
+            ->setAllowedValues('order_by', ['id', 'name', 'path', 'created_at', 'updated_at', 'last_activity_at'])
+        ;
+        $resolver->setDefined('sort')
+            ->setAllowedValues('sort', ['asc', 'desc'])
+        ;
+        $resolver->setDefined('search');
+        $resolver->setDefined('simple')
+            ->setAllowedTypes('simple', 'bool')
+            ->setNormalizer('simple', $booleanNormalizer)
+        ;
+        $resolver->setDefined('owned')
+            ->setAllowedTypes('owned', 'bool')
+            ->setNormalizer('owned', $booleanNormalizer)
+        ;
+        $resolver->setDefined('starred')
+            ->setAllowedTypes('starred', 'bool')
+            ->setNormalizer('starred', $booleanNormalizer)
+        ;
+
+        return $this->get('groups/'.$this->encodePath($id).'/projects', $resolver->resolve($parameters));
     }
 }
