@@ -1,5 +1,8 @@
 <?php namespace Gitlab\Api;
 
+use Symfony\Component\OptionsResolver\Exception\InvalidOptionsException;
+use Symfony\Component\OptionsResolver\Exception\UndefinedOptionsException;
+
 class MergeRequests extends AbstractApi
 {
     const STATE_ALL = 'all';
@@ -7,80 +10,65 @@ class MergeRequests extends AbstractApi
     const STATE_OPENED = 'opened';
     const STATE_CLOSED = 'closed';
 
-    const ORDER_BY = 'created_at';
-    const SORT = 'asc';
-
     /**
-     * @param int $project_id
-     * @param string $state
-     * @param int $page
-     * @param int $per_page
-     * @param string $order_by
-     * @param string $sort
-     * @param string $object
+     * @param int   $project_id
+     * @param array $parameters {
+     *
+     *     @var int[]              $iids           Return the request having the given iid.
+     *     @var string             $state          Return all merge requests or just those that are opened, closed, or
+     *                                             merged.
+     *     @var string             $order_by       Return requests ordered by created_at or updated_at fields. Default
+     *                                             is created_at.
+     *     @var string             $sort           Return requests sorted in asc or desc order. Default is desc.
+     *     @var string             $milestone      Return merge requests for a specific milestone.
+     *     @var string             $view           If simple, returns the iid, URL, title, description, and basic state
+     *                                             of merge request.
+     *     @var string             $labels         Return merge requests matching a comma separated list of labels.
+     *     @var \DateTimeInterface $created_after  Return merge requests created after the given time (inclusive).
+     *     @var \DateTimeInterface $created_before Return merge requests created before the given time (inclusive).
+     * }
+     *
+     * @throws UndefinedOptionsException If an option name is undefined.
+     * @throws InvalidOptionsException   If an option doesn't fulfill the specified validation rules.
+     *
      * @return mixed
      */
-    public function getList($project_id, $state = self::STATE_ALL, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = self::SORT, $object = 'merge_requests')
+    public function all($project_id, array $parameters = [])
     {
-        return $this->get($this->getProjectPath($project_id, $object), array(
-            'page' => $page,
-            'per_page' => $per_page,
-            'state' => $state,
-            'order_by' => $order_by,
-            'sort' => $sort
-        ));
-    }
+        $resolver = $this->createOptionsResolver();
+        $datetimeNormalizer = function (\DateTimeInterface $value) {
+            return $value->format('c');
+        };
+        $resolver->setDefined('iids')
+            ->setAllowedTypes('iids', 'array')
+            ->setAllowedValues('iids', function (array $value) {
+                return count($value) == count(array_filter($value, 'is_int'));
+            })
+        ;
+        $resolver->setDefined('state')
+            ->setAllowedValues('state', ['all', 'opened', 'merged', 'closed'])
+        ;
+        $resolver->setDefined('order_by')
+            ->setAllowedValues('order_by', ['created_at', 'updated_at'])
+        ;
+        $resolver->setDefined('sort')
+            ->setAllowedValues('sort', ['asc', 'desc'])
+        ;
+        $resolver->setDefined('milestone');
+        $resolver->setDefined('view')
+            ->setAllowedValues('view', ['simple'])
+        ;
+        $resolver->setDefined('labels');
+        $resolver->setDefined('created_after')
+            ->setAllowedTypes('created_after', \DateTimeInterface::class)
+            ->setNormalizer('created_after', $datetimeNormalizer)
+        ;
+        $resolver->setDefined('created_before')
+            ->setAllowedTypes('created_before', \DateTimeInterface::class)
+            ->setNormalizer('created_before', $datetimeNormalizer)
+        ;
 
-    /**
-     * @param int $project_id
-     * @param int $page
-     * @param int $per_page
-     * @param string $order_by
-     * @param string $sort
-     * @return mixed
-     */
-    public function all($project_id, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = self::SORT)
-    {
-        return $this->getList($project_id, self::STATE_ALL, $page, $per_page, $order_by, $sort);
-    }
-
-    /**
-     * @param int $project_id
-     * @param int $page
-     * @param int $per_page
-     * @param string $order_by
-     * @param string $sort
-     * @return mixed
-     */
-    public function merged($project_id, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = self::SORT)
-    {
-        return $this->getList($project_id, self::STATE_MERGED, $page, $per_page, $order_by, $sort);
-    }
-
-    /**
-     * @param int $project_id
-     * @param int $page
-     * @param int $per_page
-     * @param string $order_by
-     * @param string $sort
-     * @return mixed
-     */
-    public function opened($project_id, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = self::SORT)
-    {
-        return $this->getList($project_id, self::STATE_OPENED, $page, $per_page, $order_by, $sort);
-    }
-
-    /**
-     * @param int $project_id
-     * @param int $page
-     * @param int $per_page
-     * @param string $order_by
-     * @param string $sort
-     * @return mixed
-     */
-    public function closed($project_id, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = self::SORT)
-    {
-        return $this->getList($project_id, self::STATE_CLOSED, $page, $per_page, $order_by, $sort);
+        return $this->get($this->getProjectPath($project_id, 'merge_requests'), $resolver->resolve($parameters));
     }
 
     /**
@@ -90,7 +78,7 @@ class MergeRequests extends AbstractApi
      */
     public function show($project_id, $mr_id)
     {
-        return $this->get($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id)));
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id)));
     }
 
     /**
@@ -123,7 +111,7 @@ class MergeRequests extends AbstractApi
      */
     public function update($project_id, $mr_id, array $params)
     {
-        return $this->put($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id)), $params);
+        return $this->put($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id)), $params);
     }
 
     /**
@@ -140,17 +128,18 @@ class MergeRequests extends AbstractApi
             $params = array('merge_commit_message' => $message);
         }
 
-        return $this->put($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id).'/merge'), $params);
+        return $this->put($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/merge'), $params);
     }
 
     /**
-     * @param int $project_id
-     * @param int $mr_id
+     * @param int   $project_id
+     * @param int   $mr_id
+     *
      * @return mixed
      */
-    public function showNotes($project_id, $mr_id, $page = 1, $per_page = self::PER_PAGE, $order_by = self::ORDER_BY, $sort = 'desc')
+    public function showNotes($project_id, $mr_id)
     {
-        return $this->getList($project_id, null, $page, $per_page, $order_by, $sort, 'merge_requests/'.$this->encodePath($mr_id).'/notes');
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/notes'));
     }
 
     /**
@@ -173,7 +162,9 @@ class MergeRequests extends AbstractApi
      */
     public function showComments($project_id, $mr_id)
     {
-        return $this->get($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id).'/comments'));
+        @trigger_error(sprintf('The %s() method is deprecated since version 9.1 and will be removed in 10.0. Use the showNotes() method instead.', __METHOD__), E_USER_DEPRECATED);
+
+        return $this->showNotes($project_id, $mr_id);
     }
 
     /**
@@ -184,9 +175,9 @@ class MergeRequests extends AbstractApi
      */
     public function addComment($project_id, $mr_id, $note)
     {
-        return $this->post($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id).'/comments'), array(
-            'note' => $note
-        ));
+        @trigger_error(sprintf('The %s() method is deprecated since version 9.1 and will be removed in 10.0. Use the addNote() method instead.', __METHOD__), E_USER_DEPRECATED);
+
+        return $this->addNote($project_id, $mr_id, $note);
     }
 
     /**
@@ -196,17 +187,7 @@ class MergeRequests extends AbstractApi
      */
     public function changes($project_id, $mr_id)
     {
-        return $this->get($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id).'/changes'));
-    }
-
-    /**
-     * @param $project_id
-     * @param $mr_iid
-     * @return mixed
-     */
-    public function getByIid($project_id, $mr_iid)
-    {
-        return $this->get($this->getProjectPath($project_id, 'merge_requests'), array('iid' => $mr_iid));
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/changes'));
     }
 
     /**
@@ -216,6 +197,60 @@ class MergeRequests extends AbstractApi
      */
     public function commits($project_id, $mr_id)
     {
-        return $this->get($this->getProjectPath($project_id, 'merge_request/'.$this->encodePath($mr_id).'/commits'));
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/commits'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $mr_id
+     * @return mixed
+     */
+    public function closesIssues($project_id, $mr_id)
+    {
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($mr_id).'/closes_issues'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $mr_id
+     *
+     * @return mixed
+     */
+    public function approvals($project_id, $merge_request_iid)
+    {
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($merge_request_iid).'/approvals'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $mr_id
+     *
+     * @return mixed
+     */
+    public function approve($project_id, $merge_request_iid)
+    {
+        return $this->post($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($merge_request_iid).'/approve'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $mr_id
+     *
+     * @return mixed
+     */
+    public function unapprove($project_id, $merge_request_iid)
+    {
+        return $this->post($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($merge_request_iid).'/unapprove'));
+    }
+
+    /**
+     * @param int $project_id
+     * @param int $merge_request_iid
+     *
+     * @return mixed
+     */
+    public function awardEmoji($project_id, $merge_request_iid)
+    {
+        return $this->get($this->getProjectPath($project_id, 'merge_requests/'.$this->encodePath($merge_request_iid).'/award_emoji'));
     }
 }
