@@ -1,14 +1,18 @@
 <?php namespace Gitlab;
 
-use Buzz\Client\Curl;
-use Buzz\Client\ClientInterface;
-
 use Gitlab\Api\AbstractApi;
 use Gitlab\Exception\InvalidArgumentException;
-use Gitlab\HttpClient\HttpClient;
-use Gitlab\HttpClient\HttpClientInterface;
-use Gitlab\HttpClient\Listener\AuthListener;
-use Gitlab\HttpClient\Listener\PaginationListener;
+use Gitlab\HttpClient\Builder;
+use Gitlab\HttpClient\Plugin\ApiVersion;
+use Gitlab\HttpClient\Plugin\History;
+use Gitlab\HttpClient\Plugin\Authentication;
+use Gitlab\HttpClient\Plugin\GitlabExceptionThrower;
+use Http\Client\Common\HttpMethodsClient;
+use Http\Client\Common\Plugin\AddHostPlugin;
+use Http\Client\Common\Plugin\HeaderDefaultsPlugin;
+use Http\Client\Common\Plugin\HistoryPlugin;
+use Http\Client\HttpClient;
+use Http\Discovery\UriFactoryDiscovery;
 
 /**
  * Simple API wrapper for Gitlab
@@ -17,6 +21,7 @@ use Gitlab\HttpClient\Listener\PaginationListener;
  *
  * @property-read \Gitlab\Api\Groups $groups
  * @property-read \Gitlab\Api\Issues $issues
+ * @property-read \Gitlab\Api\Jobs $jobs
  * @property-read \Gitlab\Api\MergeRequests $merge_requests
  * @property-read \Gitlab\Api\MergeRequests $mr
  * @property-read \Gitlab\Api\Milestones $milestones
@@ -54,48 +59,190 @@ class Client
     const AUTH_OAUTH_TOKEN = 'oauth_token';
 
     /**
-     * @var array
+     * @var History
      */
-    private $options = array(
-        'user_agent'  => 'php-gitlab-api (http://github.com/m4tthumphrey/php-gitlab-api)',
-        'timeout'     => 60
-    );
-
-    private $baseUrl;
+    private $responseHistory;
 
     /**
-     * The Buzz instance used to communicate with Gitlab
-     *
-     * @var HttpClient
+     * @var Builder
      */
-    private $httpClient;
+    private $httpClientBuilder;
 
     /**
      * Instantiate a new Gitlab client
      *
-     * @param string               $baseUrl
-     * @param null|ClientInterface $httpClient Buzz client
-     * @param array                $options
+     * @param Builder $httpClientBuilder
      */
-    public function __construct($baseUrl, ClientInterface $httpClient = null, array $options = array())
+    public function __construct(Builder $httpClientBuilder = null)
     {
-        foreach ($options as $name => $value) {
-            $this->setOption($name, $value);
-        }
+        $this->responseHistory = new History();
+        $this->httpClientBuilder = $httpClientBuilder ?: new Builder();
 
-        $httpClient = $httpClient ?: new Curl();
-        $httpClient->setTimeout($this->options['timeout']);
-        $httpClient->setVerifyPeer(false);
+        $this->httpClientBuilder->addPlugin(new GitlabExceptionThrower());
+        $this->httpClientBuilder->addPlugin(new HistoryPlugin($this->responseHistory));
+        $this->httpClientBuilder->addPlugin(new ApiVersion());
+        $this->httpClientBuilder->addPlugin(new HeaderDefaultsPlugin([
+            'User-Agent' => 'php-gitlab-api (http://github.com/m4tthumphrey/php-gitlab-api)',
+        ]));
 
-        $this->baseUrl     = $baseUrl;
-        $this->httpClient  = new HttpClient($this->baseUrl, $this->options, $httpClient);
+        $this->setUrl('https://gitlab.com');
+    }
 
-        /**
-         * a Pagination listener on Response
-         */
-        $this->httpClient->addListener(
-            new PaginationListener()
-        );
+    /**
+     * Create a Gitlab\Client using an url.
+     *
+     * @param string $url
+     *
+     * @return Client
+     */
+    public static function create($url)
+    {
+        $client = new self();
+        $client->setUrl($url);
+
+        return $client;
+    }
+
+    /**
+     * Create a Gitlab\Client using an HttpClient.
+     *
+     * @param HttpClient $httpClient
+     *
+     * @return Client
+     */
+    public static function createWithHttpClient(HttpClient $httpClient)
+    {
+        $builder = new Builder($httpClient);
+
+        return new self($builder);
+    }
+
+    /**
+     * @return Api\DeployKeys
+     */
+    public function deployKeys()
+    {
+        return new Api\DeployKeys($this);
+    }
+
+    /**
+     * @return Api\Groups
+     */
+    public function groups()
+    {
+        return new Api\Groups($this);
+    }
+
+    /**
+     * @return Api\Issues
+     */
+    public function issues()
+    {
+        return new Api\Issues($this);
+    }
+
+    /**
+     * @return Api\IssueBoards
+     */
+    public function issueBoards()
+    {
+        return new Api\IssueBoards($this);
+    }
+
+    /**
+     * @return Api\Jobs
+     */
+    public function jobs()
+    {
+        return new Api\Jobs($this);
+    }
+
+    /**
+     * @return Api\MergeRequests
+     */
+    public function mergeRequests()
+    {
+        return new Api\MergeRequests($this);
+    }
+
+    /**
+     * @return Api\Milestones
+     */
+    public function milestones()
+    {
+        return new Api\Milestones($this);
+    }
+
+    /**
+     * @return Api\ProjectNamespaces
+     */
+    public function namespaces()
+    {
+        return new Api\ProjectNamespaces($this);
+    }
+
+    /**
+     * @return Api\Projects
+     */
+    public function projects()
+    {
+        return new Api\Projects($this);
+    }
+
+    /**
+     * @return Api\Repositories
+     */
+    public function repositories()
+    {
+        return new Api\Repositories($this);
+    }
+
+    /**
+     * @return Api\RepositoryFiles
+     */
+    public function repositoryFiles()
+    {
+        return new Api\RepositoryFiles($this);
+    }
+
+    /**
+     * @return Api\Snippets
+     */
+    public function snippets()
+    {
+        return new Api\Snippets($this);
+    }
+
+    /**
+     * @return Api\SystemHooks
+     */
+    public function systemHooks()
+    {
+        return new Api\SystemHooks($this);
+    }
+
+    /**
+     * @return Api\Users
+     */
+    public function users()
+    {
+        return new Api\Users($this);
+    }
+
+    /**
+     * @return Api\Keys
+     */
+    public function keys()
+    {
+        return new Api\Keys($this);
+    }
+
+    /**
+     * @return Api\Tags
+     */
+    public function tags()
+    {
+        return new Api\Tags($this);
     }
 
     /**
@@ -109,68 +256,58 @@ class Client
         switch ($name) {
 
             case 'deploy_keys':
-                $api = new Api\DeployKeys($this);
-                break;
+                return $this->deployKeys();
 
             case 'groups':
-                $api = new Api\Groups($this);
-                break;
+                return $this->groups();
 
             case 'issues':
-                $api = new Api\Issues($this);
-                break;
+                return $this->issues();
+
+            case 'board':
+            case 'issue_boards':
+                return $this->issueBoards();
+            case 'jobs':
+                return $this->jobs();
 
             case 'mr':
             case 'merge_requests':
-                $api = new Api\MergeRequests($this);
-                break;
+                return $this->mergeRequests();
 
             case 'milestones':
             case 'ms':
-                $api = new Api\Milestones($this);
-                break;
+                return $this->milestones();
 
             case 'namespaces':
             case 'ns':
-                $api = new Api\ProjectNamespaces($this);
-                break;
+                return $this->namespaces();
 
             case 'projects':
-                $api = new Api\Projects($this);
-                break;
+                return $this->projects();
 
             case 'repo':
             case 'repositories':
-                $api = new Api\Repositories($this);
-                break;
+                return $this->repositories();
 
             case 'snippets':
-                $api = new Api\Snippets($this);
-                break;
+                return $this->snippets();
 
             case 'hooks':
             case 'system_hooks':
-                $api = new Api\SystemHooks($this);
-                break;
+                return $this->systemHooks();
 
             case 'users':
-                $api = new Api\Users($this);
-                break;
+                return $this->users();
 
             case 'keys':
-                $api = new Api\Keys($this);
-                break;
+                return $this->keys();
 
             case 'tags':
-                $api = new Api\Tags($this);
-                break;
+                return $this->tags();
 
             default:
                 throw new InvalidArgumentException('Invalid endpoint: "'.$name.'"');
-
         }
-
-        return $api;
     }
 
     /**
@@ -183,107 +320,21 @@ class Client
      */
     public function authenticate($token, $authMethod = self::AUTH_URL_TOKEN, $sudo = null)
     {
-        $this->httpClient->addListener(
-            new AuthListener(
-                $authMethod,
-                $token,
-                $sudo
-            )
-        );
-
-        return $this;
-    }
-
-    /**
-     * @return HttpClient
-     */
-    public function getHttpClient()
-    {
-        return $this->httpClient;
-    }
-
-    /**
-     * @param HttpClientInterface $httpClient
-     * @return $this
-     */
-    public function setHttpClient(HttpClientInterface $httpClient)
-    {
-        $this->httpClient = $httpClient;
+        $this->httpClientBuilder->removePlugin(Authentication::class);
+        $this->httpClientBuilder->addPlugin(new Authentication($authMethod, $token, $sudo));
 
         return $this;
     }
 
     /**
      * @param string $url
-     * @return $this
-     */
-    public function setBaseUrl($url)
-    {
-        $this->baseUrl = $url;
-
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getBaseUrl()
-    {
-        return $this->baseUrl;
-    }
-
-    /**
-     * Clears used headers
      *
      * @return $this
      */
-    public function clearHeaders()
+    public function setUrl($url)
     {
-        $this->httpClient->clearHeaders();
-
-        return $this;
-    }
-
-    /**
-     * @param array $headers
-     * @return $this
-     */
-    public function setHeaders(array $headers)
-    {
-        $this->httpClient->setHeaders($headers);
-
-        return $this;
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return mixed
-     *
-     * @throws InvalidArgumentException
-     */
-    public function getOption($name)
-    {
-        if (!array_key_exists($name, $this->options)) {
-            throw new InvalidArgumentException(sprintf('Undefined option called: "%s"', $name));
-        }
-
-        return $this->options[$name];
-    }
-
-    /**
-     * @param string $name
-     * @param mixed $value
-     * @throws InvalidArgumentException
-     * @return $this
-     */
-    public function setOption($name, $value)
-    {
-        if (!array_key_exists($name, $this->options)) {
-            throw new InvalidArgumentException(sprintf('Undefined option called: "%s"', $name));
-        }
-
-        $this->options[$name] = $value;
+        $this->httpClientBuilder->removePlugin(AddHostPlugin::class);
+        $this->httpClientBuilder->addPlugin(new AddHostPlugin(UriFactoryDiscovery::find()->createUri($url)));
 
         return $this;
     }
@@ -295,5 +346,21 @@ class Client
     public function __get($api)
     {
         return $this->api($api);
+    }
+
+    /**
+     * @return HttpMethodsClient
+     */
+    public function getHttpClient()
+    {
+        return $this->httpClientBuilder->getHttpClient();
+    }
+
+    /**
+     * @return History
+     */
+    public function getResponseHistory()
+    {
+        return $this->responseHistory;
     }
 }
