@@ -1,7 +1,7 @@
 <?php namespace Gitlab\Model;
 
+use Gitlab\Api\Projects;
 use Gitlab\Client;
-use Gitlab\Api\AbstractApi as Api;
 
 /**
  * Class Branch
@@ -57,7 +57,7 @@ class Branch extends AbstractModel
      */
     public function show()
     {
-        $data = $this->api('repositories')->branch($this->project->id, $this->name);
+        $data = $this->client->repositories()->branch($this->project->id, $this->name);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -69,7 +69,7 @@ class Branch extends AbstractModel
      */
     public function protect($devPush = false, $devMerge = false)
     {
-        $data = $this->api('repositories')->protectBranch($this->project->id, $this->name, $devPush, $devMerge);
+        $data = $this->client->repositories()->protectBranch($this->project->id, $this->name, $devPush, $devMerge);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -79,7 +79,7 @@ class Branch extends AbstractModel
      */
     public function unprotect()
     {
-        $data = $this->api('repositories')->unprotectBranch($this->project->id, $this->name);
+        $data = $this->client->repositories()->unprotectBranch($this->project->id, $this->name);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -89,19 +89,21 @@ class Branch extends AbstractModel
      */
     public function delete()
     {
-        $this->api('repositories')->deleteBranch($this->project->id, $this->name);
+        $this->client->repositories()->deleteBranch($this->project->id, $this->name);
 
         return true;
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Projects::commits for available parameters.
+     *
      * @return Commit[]
      */
-    public function commits($page = 1, $per_page = Api::PER_PAGE)
+    public function commits(array $parameters = [])
     {
-        return $this->project->commits($page, $per_page, $this->name);
+        return $this->project->commits($parameters);
     }
 
     /**
@@ -112,7 +114,7 @@ class Branch extends AbstractModel
      */
     public function createFile($file_path, $content, $commit_message)
     {
-        $data = $this->api('repositories')->createFile($this->project->id, $file_path, $content, $this->name, $commit_message);
+        $data = $this->client->repositories()->createFile($this->project->id, $file_path, $content, $this->name, $commit_message);
 
         return File::fromArray($this->getClient(), $this->project, $data);
     }
@@ -125,7 +127,7 @@ class Branch extends AbstractModel
      */
     public function updateFile($file_path, $content, $commit_message)
     {
-        $data = $this->api('repositories')->updateFile($this->project->id, $file_path, $content, $this->name, $commit_message);
+        $data = $this->client->repositories()->updateFile($this->project->id, $file_path, $content, $this->name, $commit_message);
 
         return File::fromArray($this->getClient(), $this->project, $data);
     }
@@ -137,7 +139,7 @@ class Branch extends AbstractModel
      */
     public function deleteFile($file_path, $commit_message)
     {
-        $this->api('repositories')->deleteFile($this->project->id, $file_path, $this->name, $commit_message);
+        $this->client->repositories()->deleteFile($this->project->id, $file_path, $this->name, $commit_message);
 
         return true;
     }
