@@ -11,6 +11,7 @@ use Gitlab\Client;
  * @property-read string $title
  * @property-read string $description
  * @property-read string $due_date
+ * @property-read string $start_date
  * @property-read string $state
  * @property-read bool $closed
  * @property-read string $updated_at
@@ -30,6 +31,7 @@ class Milestone extends AbstractModel
         'title',
         'description',
         'due_date',
+        'start_date',
         'state',
         'closed',
         'updated_at',
@@ -66,7 +68,7 @@ class Milestone extends AbstractModel
      */
     public function show()
     {
-        $data = $this->api('milestones')->show($this->project->id, $this->id);
+        $data = $this->client->milestones()->show($this->project->id, $this->id);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -77,7 +79,7 @@ class Milestone extends AbstractModel
      */
     public function update(array $params)
     {
-        $data = $this->api('milestones')->update($this->project->id, $this->id, $params);
+        $data = $this->client->milestones()->update($this->project->id, $this->id, $params);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -103,7 +105,7 @@ class Milestone extends AbstractModel
      */
     public function issues()
     {
-        $data = $this->api('milestones')->issues($this->project->id, $this->id);
+        $data = $this->client->milestones()->issues($this->project->id, $this->id);
 
         $issues = array();
         foreach ($data as $issue) {
