@@ -1,8 +1,8 @@
 <?php namespace Gitlab\Model;
 
-use Gitlab\Api\MergeRequests;
+use Gitlab\Api\Projects;
+use Gitlab\Api\Repositories;
 use Gitlab\Client;
-use Gitlab\Api\AbstractApi as Api;
 
 /**
  * Class Project
@@ -93,7 +93,7 @@ class Project extends AbstractModel
      */
     public static function create(Client $client, $name, array $params = array())
     {
-        $data = $client->api('projects')->create($name, $params);
+        $data = $client->projects()->create($name, $params);
 
         return static::fromArray($client, $data);
     }
@@ -107,7 +107,7 @@ class Project extends AbstractModel
      */
     public static function createForUser($user_id, Client $client, $name, array $params = array())
     {
-        $data = $client->api('projects')->createForUser($user_id, $name, $params);
+        $data = $client->projects()->createForUser($user_id, $name, $params);
 
         return static::fromArray($client, $data);
     }
@@ -126,7 +126,7 @@ class Project extends AbstractModel
      */
     public function show()
     {
-        $data = $this->api('projects')->show($this->id);
+        $data = $this->client->projects()->show($this->id);
 
         return static::fromArray($this->getClient(), $data);
     }
@@ -137,7 +137,7 @@ class Project extends AbstractModel
      */
     public function update(array $params)
     {
-        $data = $this->api('projects')->update($this->id, $params);
+        $data = $this->client->projects()->update($this->id, $params);
 
         return static::fromArray($this->getClient(), $data);
     }
@@ -147,7 +147,7 @@ class Project extends AbstractModel
      */
     public function archive()
     {
-        $data = $this->api("projects")->archive($this->id);
+        $data = $this->client->projects()->archive($this->id);
 
         return static::fromArray($this->getClient(), $data);
     }
@@ -157,7 +157,7 @@ class Project extends AbstractModel
      */
     public function unarchive()
     {
-        $data = $this->api("projects")->unarchive($this->id);
+        $data = $this->client->projects()->unarchive($this->id);
 
         return static::fromArray($this->getClient(), $data);
     }
@@ -167,7 +167,7 @@ class Project extends AbstractModel
      */
     public function remove()
     {
-        $this->api('projects')->remove($this->id);
+        $this->client->projects()->remove($this->id);
 
         return true;
     }
@@ -178,7 +178,7 @@ class Project extends AbstractModel
      */
     public function members($username_query = null)
     {
-        $data = $this->api('projects')->members($this->id, $username_query);
+        $data = $this->client->projects()->members($this->id, $username_query);
 
         $members = array();
         foreach ($data as $member) {
@@ -194,7 +194,7 @@ class Project extends AbstractModel
      */
     public function member($user_id)
     {
-        $data = $this->api('projects')->member($this->id, $user_id);
+        $data = $this->client->projects()->member($this->id, $user_id);
 
         return User::fromArray($this->getClient(), $data);
     }
@@ -206,7 +206,7 @@ class Project extends AbstractModel
      */
     public function addMember($user_id, $access_level)
     {
-        $data = $this->api('projects')->addMember($this->id, $user_id, $access_level);
+        $data = $this->client->projects()->addMember($this->id, $user_id, $access_level);
 
         return User::fromArray($this->getClient(), $data);
     }
@@ -218,7 +218,7 @@ class Project extends AbstractModel
      */
     public function saveMember($user_id, $access_level)
     {
-        $data = $this->api('projects')->saveMember($this->id, $user_id, $access_level);
+        $data = $this->client->projects()->saveMember($this->id, $user_id, $access_level);
 
         return User::fromArray($this->getClient(), $data);
     }
@@ -229,19 +229,21 @@ class Project extends AbstractModel
      */
     public function removeMember($user_id)
     {
-        $this->api('projects')->removeMember($this->id, $user_id);
+        $this->client->projects()->removeMember($this->id, $user_id);
 
         return true;
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Projects::hooks() for available parameters.
+     *
      * @return ProjectHook[]
      */
-    public function hooks($page = 1, $per_page = Api::PER_PAGE)
+    public function hooks(array $parameters = [])
     {
-        $data = $this->api('projects')->hooks($this->id, $page, $per_page);
+        $data = $this->client->projects()->hooks($this->id, $parameters);
 
         $hooks = array();
         foreach ($data as $hook) {
@@ -269,7 +271,7 @@ class Project extends AbstractModel
      */
     public function addHook($url, array $events = array())
     {
-        $data = $this->api('projects')->addHook($this->id, $url, $events);
+        $data = $this->client->projects()->addHook($this->id, $url, $events);
 
         return ProjectHook::fromArray($this->getClient(), $this, $data);
     }
@@ -300,9 +302,9 @@ class Project extends AbstractModel
     /**
      * @return Key[]
      */
-    public function keys()
+    public function deployKeys()
     {
-        $data = $this->api('projects')->keys($this->id);
+        $data = $this->client->projects()->deployKeys($this->id);
 
         $keys = array();
         foreach ($data as $key) {
@@ -316,9 +318,9 @@ class Project extends AbstractModel
      * @param int $key_id
      * @return Key
      */
-    public function key($key_id)
+    public function deployKey($key_id)
     {
-        $data = $this->api('projects')->key($this->id, $key_id);
+        $data = $this->client->projects()->deployKey($this->id, $key_id);
 
         return Key::fromArray($this->getClient(), $data);
     }
@@ -328,9 +330,9 @@ class Project extends AbstractModel
      * @param string $key
      * @return Key
      */
-    public function addKey($title, $key)
+    public function addDeployKey($title, $key)
     {
-        $data = $this->api('projects')->addKey($this->id, $title, $key);
+        $data = $this->client->projects()->addDeployKey($this->id, $title, $key);
 
         return Key::fromArray($this->getClient(), $data);
     }
@@ -339,9 +341,9 @@ class Project extends AbstractModel
      * @param string $key_id
      * @return bool
      */
-    public function removeKey($key_id)
+    public function deleteDeployKey($key_id)
     {
-        $this->api('projects')->removeKey($this->id, $key_id);
+        $this->client->projects()->deleteDeployKey($this->id, $key_id);
 
         return true;
     }
@@ -350,20 +352,9 @@ class Project extends AbstractModel
      * @param string $key_id
      * @return bool
      */
-    public function enableKey($key_id)
+    public function enableDeployKey($key_id)
     {
-        $this->api('projects')->enableKey($this->id, $key_id);
-
-        return true;
-    }
-
-    /**
-     * @param string $key_id
-     * @return bool
-     */
-    public function disableKey($key_id)
-    {
-        $this->api('projects')->disableKey($this->id, $key_id);
+        $this->client->projects()->enableDeployKey($this->id, $key_id);
 
         return true;
     }
@@ -375,7 +366,7 @@ class Project extends AbstractModel
      */
     public function createBranch($name, $ref)
     {
-        $data = $this->api('repositories')->createBranch($this->id, $name, $ref);
+        $data = $this->client->repositories()->createBranch($this->id, $name, $ref);
 
         return Branch::fromArray($this->getClient(), $this, $data);
     }
@@ -386,7 +377,7 @@ class Project extends AbstractModel
      */
     public function deleteBranch($name)
     {
-        $this->api('repositories')->deleteBranch($this->id, $name);
+        $this->client->repositories()->deleteBranch($this->id, $name);
 
         return true;
     }
@@ -396,7 +387,7 @@ class Project extends AbstractModel
      */
     public function branches()
     {
-        $data = $this->api('repo')->branches($this->id);
+        $data = $this->client->repositories()->branches($this->id);
 
         $branches = array();
         foreach ($data as $branch) {
@@ -449,7 +440,7 @@ class Project extends AbstractModel
      */
     public function tags()
     {
-        $data = $this->api('repo')->tags($this->id);
+        $data = $this->client->repositories()->tags($this->id);
 
         $tags = array();
         foreach ($data as $tag) {
@@ -460,14 +451,15 @@ class Project extends AbstractModel
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
-     * @param string $ref_name
+     * @param array $parameters
+     *
+     * @see Repositories::commits() for available parameters.
+     *
      * @return Commit[]
      */
-    public function commits($page = 0, $per_page = Api::PER_PAGE, $ref_name = null)
+    public function commits(array $parameters = [])
     {
-        $data = $this->api('repo')->commits($this->id, $page, $per_page, $ref_name);
+        $data = $this->client->repositories()->commits($this->id, $parameters);
 
         $commits = array();
         foreach ($data as $commit) {
@@ -483,20 +475,22 @@ class Project extends AbstractModel
      */
     public function commit($sha)
     {
-        $data = $this->api('repo')->commit($this->id, $sha);
+        $data = $this->client->repositories()->commit($this->id, $sha);
 
         return Commit::fromArray($this->getClient(), $this, $data);
     }
 
     /**
      * @param string $ref
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Repositories::commitComments() for available parameters.
+     *
      * @return Commit[]
      */
-    public function commitComments($ref, $page = 0, $per_page = Api::PER_PAGE)
+    public function commitComments($ref, array $parameters = [])
     {
-        $data = $this->api('repo')->commitComments($this->id, $ref, $page, $per_page);
+        $data = $this->client->repositories()->commitComments($this->id, $ref, $parameters);
 
         $comments = array();
         foreach ($data as $comment) {
@@ -514,7 +508,7 @@ class Project extends AbstractModel
      */
     public function createCommitComment($ref, $note, array $params = array())
     {
-        $data = $this->api('repo')->createCommitComment($this->id, $ref, $note, $params);
+        $data = $this->client->repositories()->createCommitComment($this->id, $ref, $note, $params);
 
         return CommitNote::fromArray($this->getClient(), $data);
     }
@@ -525,7 +519,7 @@ class Project extends AbstractModel
      */
     public function diff($sha)
     {
-        return $this->api('repo')->diff($this->id, $sha);
+        return $this->client->repositories()->diff($this->id, $sha);
     }
 
     /**
@@ -535,7 +529,7 @@ class Project extends AbstractModel
      */
     public function compare($from, $to)
     {
-        $data = $this->api('repo')->compare($this->id, $from, $to);
+        $data = $this->client->repositories()->compare($this->id, $from, $to);
 
         return Comparison::fromArray($this->getClient(), $this, $data);
     }
@@ -546,7 +540,7 @@ class Project extends AbstractModel
      */
     public function tree(array $params = array())
     {
-        $data = $this->api('repo')->tree($this->id, $params);
+        $data = $this->client->repositories()->tree($this->id, $params);
 
         $tree = array();
         foreach ($data as $node) {
@@ -563,7 +557,7 @@ class Project extends AbstractModel
      */
     public function blob($sha, $filepath)
     {
-        return $this->api('repo')->blob($this->id, $sha, $filepath);
+        return $this->client->repositories()->blob($this->id, $sha, $filepath);
     }
 
     /**
@@ -574,7 +568,7 @@ class Project extends AbstractModel
      */
     public function getFile($sha, $filepath)
     {
-        return $this->api('repo')->getFile($this->id, $filepath, $sha);
+        return $this->client->repositories()->getFile($this->id, $filepath, $sha);
     }
 
     /**
@@ -588,7 +582,22 @@ class Project extends AbstractModel
      */
     public function createFile($file_path, $content, $branch_name, $commit_message, $author_email = null, $author_name = null)
     {
-        $data = $this->api('repo')->createFile($this->id, $file_path, $content, $branch_name, $commit_message, null, $author_email, $author_name);
+        $parameters = [
+            'file_path' => $file_path,
+            'branch' => $branch_name,
+            'content' => $content,
+            'commit_message' => $commit_message,
+        ];
+
+        if ($author_email !== null) {
+            $parameters['author_email'] = $author_email;
+        }
+
+        if ($author_name !== null) {
+            $parameters['author_name'] = $author_name;
+        }
+
+        $data = $this->client->repositoryFiles()->createFile($this->id, $parameters);
 
         return File::fromArray($this->getClient(), $this, $data);
     }
@@ -604,7 +613,22 @@ class Project extends AbstractModel
      */
     public function updateFile($file_path, $content, $branch_name, $commit_message, $author_email = null, $author_name = null)
     {
-        $data = $this->api('repo')->updateFile($this->id, $file_path, $content, $branch_name, $commit_message, null, $author_email, $author_name);
+        $parameters = [
+            'file_path' => $file_path,
+            'branch' => $branch_name,
+            'content' => $content,
+            'commit_message' => $commit_message,
+        ];
+
+        if ($author_email !== null) {
+            $parameters['author_email'] = $author_email;
+        }
+
+        if ($author_name !== null) {
+            $parameters['author_name'] = $author_name;
+        }
+
+        $data = $this->client->repositoryFiles()->updateFile($this->id, $parameters);
 
         return File::fromArray($this->getClient(), $this, $data);
     }
@@ -619,19 +643,35 @@ class Project extends AbstractModel
      */
     public function deleteFile($file_path, $branch_name, $commit_message, $author_email = null, $author_name = null)
     {
-        $this->api('repo')->deleteFile($this->id, $file_path, $branch_name, $commit_message, $author_email, $author_name);
+        $parameters = [
+            'file_path' => $file_path,
+            'branch' => $branch_name,
+            'commit_message' => $commit_message,
+        ];
+
+        if ($author_email !== null) {
+            $parameters['author_email'] = $author_email;
+        }
+
+        if ($author_name !== null) {
+            $parameters['author_name'] = $author_name;
+        }
+
+        $this->client->repositoryFiles()->deleteFile($this->id, $parameters);
 
         return true;
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Projects::events() for available parameters.
+     *
      * @return Event[]
      */
-    public function events($page = 1, $per_page = Api::PER_PAGE)
+    public function events(array $parameters = [])
     {
-        $data = $this->api('projects')->events($this->id, $page, $per_page);
+        $data = $this->client->projects()->events($this->id, $parameters);
 
         $events = array();
         foreach ($data as $event) {
@@ -642,14 +682,15 @@ class Project extends AbstractModel
     }
 
     /**
-     * @param int    $page
-     * @param int    $per_page
-     * @param string $state
+     * @param array $parameters
+     *
+     * @see MergeRequests::all() for available parameters.
+     *
      * @return MergeRequest[]
      */
-    public function mergeRequests($page = 1, $per_page = Api::PER_PAGE, $state = MergeRequests::STATE_ALL)
+    public function mergeRequests(array $parameters = [])
     {
-        $data = $this->api('mr')->$state($this->id, $page, $per_page);
+        $data = $this->client->mergeRequests()->all($this->id, $parameters);
 
         $mrs = array();
         foreach ($data as $mr) {
@@ -680,7 +721,7 @@ class Project extends AbstractModel
      */
     public function createMergeRequest($source, $target, $title, $assignee = null, $description = null)
     {
-        $data = $this->api('mr')->create($this->id, $source, $target, $title, $assignee, null, $description);
+        $data = $this->client->mergeRequests()->create($this->id, $source, $target, $title, $assignee, null, $description);
 
         return MergeRequest::fromArray($this->getClient(), $this, $data);
     }
@@ -731,13 +772,15 @@ class Project extends AbstractModel
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Issues::all() for available parameters.
+     *
      * @return Issue[]
      */
-    public function issues($page = 1, $per_page = Api::PER_PAGE)
+    public function issues(array $parameters = [])
     {
-        $data = $this->api('issues')->all($this->id, $page, $per_page);
+        $data = $this->client->issues()->all($this->id, $parameters);
 
         $issues = array();
         foreach ($data as $issue) {
@@ -755,7 +798,7 @@ class Project extends AbstractModel
     public function createIssue($title, array $params = array())
     {
         $params['title'] = $title;
-        $data = $this->api('issues')->create($this->id, $params);
+        $data = $this->client->issues()->create($this->id, $params);
 
         return Issue::fromArray($this->getClient(), $this, $data);
     }
@@ -807,13 +850,15 @@ class Project extends AbstractModel
     }
 
     /**
-     * @param int $page
-     * @param int $per_page
+     * @param array $parameters
+     *
+     * @see Milestones::all() for available parameters.
+     *
      * @return Milestone[]
      */
-    public function milestones($page = 1, $per_page = Api::PER_PAGE)
+    public function milestones(array $parameters = [])
     {
-        $data = $this->api('milestones')->all($this->id, $page, $per_page);
+        $data = $this->client->milestones()->all($this->id, $parameters);
 
         $milestones = array();
         foreach ($data as $milestone) {
@@ -831,7 +876,7 @@ class Project extends AbstractModel
     public function createMilestone($title, array $params = array())
     {
         $params['title'] = $title;
-        $data = $this->api('milestones')->create($this->id, $params);
+        $data = $this->client->milestones()->create($this->id, $params);
 
         return Milestone::fromArray($this->getClient(), $this, $data);
     }
@@ -875,7 +920,7 @@ class Project extends AbstractModel
      */
     public function snippets()
     {
-        $data = $this->api('snippets')->all($this->id);
+        $data = $this->client->snippets()->all($this->id);
 
         $snippets = array();
         foreach ($data as $snippet) {
@@ -889,12 +934,11 @@ class Project extends AbstractModel
      * @param string $title
      * @param string $filename
      * @param string $code
-     * @param string $lifetime
      * @return Snippet
      */
-    public function createSnippet($title, $filename, $code, $lifetime = null)
+    public function createSnippet($title, $filename, $code)
     {
-        $data = $this->api('snippets')->create($this->id, $title, $filename, $code, $lifetime);
+        $data = $this->client->snippets()->create($this->id, $title, $filename, $code);
 
         return Snippet::fromArray($this->getClient(), $this, $data);
     }
@@ -912,7 +956,7 @@ class Project extends AbstractModel
 
     /**
      * @param int $id
-     * @return Snippet
+     * @return string
      */
     public function snippetContent($id)
     {
@@ -961,7 +1005,7 @@ class Project extends AbstractModel
      */
     public function forkTo($id)
     {
-        $data = $this->api('projects')->createForkRelation($id, $this->id);
+        $data = $this->client->projects()->createForkRelation($id, $this->id);
 
         return Project::fromArray($this->getClient(), $data);
     }
@@ -981,7 +1025,7 @@ class Project extends AbstractModel
      */
     public function createForkRelation($id)
     {
-        $data = $this->api('projects')->createForkRelation($this->id, $id);
+        $data = $this->client->projects()->createForkRelation($this->id, $id);
 
         return Project::fromArray($this->getClient(), $data);
     }
@@ -991,7 +1035,7 @@ class Project extends AbstractModel
      */
     public function removeForkRelation()
     {
-        $this->api('projects')->removeForkRelation($this->id);
+        $this->client->projects()->removeForkRelation($this->id);
 
         return true;
     }
@@ -1003,7 +1047,7 @@ class Project extends AbstractModel
      */
     public function setService($service_name, array $params = array())
     {
-        $this->api('projects')->setService($this->id, $service_name, $params);
+        $this->client->projects()->setService($this->id, $service_name, $params);
 
         return true;
     }
@@ -1014,7 +1058,7 @@ class Project extends AbstractModel
      */
     public function removeService($service_name)
     {
-        $this->api('projects')->removeService($this->id, $service_name);
+        $this->client->projects()->removeService($this->id, $service_name);
 
         return true;
     }
@@ -1024,7 +1068,7 @@ class Project extends AbstractModel
      */
     public function labels()
     {
-        $data = $this->api('projects')->labels($this->id);
+        $data = $this->client->projects()->labels($this->id);
 
         $labels = array();
         foreach ($data as $label) {
@@ -1041,7 +1085,7 @@ class Project extends AbstractModel
      */
     public function addLabel($name, $color)
     {
-        $data = $this->api('projects')->addLabel($this->id, array(
+        $data = $this->client->projects()->addLabel($this->id, array(
             'name' => $name,
             'color' => $color
         ));
@@ -1062,7 +1106,7 @@ class Project extends AbstractModel
 
         $params['name'] = $name;
 
-        $data = $this->api('projects')->updateLabel($this->id, $params);
+        $data = $this->client->projects()->updateLabel($this->id, $params);
 
         return Label::fromArray($this->getClient(), $this, $data);
     }
@@ -1073,7 +1117,7 @@ class Project extends AbstractModel
      */
     public function removeLabel($name)
     {
-        $this->api('projects')->removeLabel($this->id, $name);
+        $this->client->projects()->removeLabel($this->id, $name);
 
         return true;
     }
@@ -1083,7 +1127,7 @@ class Project extends AbstractModel
      */
     public function contributors()
     {
-        $data = $this->api('repo')->contributors($this->id);
+        $data = $this->client->repositories()->contributors($this->id);
 
         $contributors = array();
         foreach ($data as $contributor) {
@@ -1091,5 +1135,49 @@ class Project extends AbstractModel
         }
 
         return $contributors;
+    }
+
+    /**
+     * @param array $scopes
+     * @return Job[]
+     */
+    public function jobs(array $scopes = [])
+    {
+        $data = $this->client->jobs()->all($this->id, $scopes);
+
+        $jobs = array();
+        foreach ($data as $job) {
+            $jobs[] = Job::fromArray($this->getClient(), $this, $job);
+        }
+
+        return $jobs;
+    }
+
+    /**
+     * @param int $pipeline_id
+     * @param array $scopes
+     * @return Job[]
+     */
+    public function pipelineJobs($pipeline_id, array $scopes = [])
+    {
+        $data = $this->client->jobs()->pipelineJobs($this->id, $pipeline_id, $scopes);
+
+        $jobs = array();
+        foreach ($data as $job) {
+            $jobs[] = Job::fromArray($this->getClient(), $this, $job);
+        }
+
+        return $jobs;
+    }
+
+    /**
+     * @param int $job_id
+     * @return Job
+     */
+    public function job($job_id)
+    {
+        $data = $this->client->jobs()->show($this->id, $job_id);
+
+        return Job::fromArray($this->getClient(), $this, $data);
     }
 }
