@@ -11,8 +11,9 @@ use Gitlab\Client;
  * @property-read bool $push_events
  * @property-read bool $issues_events
  * @property-read bool $merge_requests_events
- * @property-read bool $build_events
+ * @property-read bool $job_events
  * @property-read bool $tag_push_events
+ * @property-read bool $pipeline_events
  * @property-read string $created_at
  * @property-read Project $project
  */
@@ -29,8 +30,9 @@ class ProjectHook extends AbstractModel
         'push_events',
         'issues_events',
         'merge_requests_events',
-        'build_events',
+        'job_events',
         'tag_push_events',
+        'pipeline_events',
         'created_at'
     );
 
@@ -64,7 +66,7 @@ class ProjectHook extends AbstractModel
      */
     public function show()
     {
-        $data = $this->api('projects')->hook($this->project->id, $this->id);
+        $data = $this->client->projects()->hook($this->project->id, $this->id);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
@@ -74,7 +76,7 @@ class ProjectHook extends AbstractModel
      */
     public function delete()
     {
-        $this->api('projects')->removeHook($this->project->id, $this->id);
+        $this->client->projects()->removeHook($this->project->id, $this->id);
 
         return true;
     }
@@ -93,7 +95,7 @@ class ProjectHook extends AbstractModel
      */
     public function update(array $params)
     {
-        $data = $this->api('projects')->updateHook($this->project->id, $this->id, $params);
+        $data = $this->client->projects()->updateHook($this->project->id, $this->id, $params);
 
         return static::fromArray($this->getClient(), $this->project, $data);
     }
