@@ -1,6 +1,7 @@
 <?php namespace Gitlab;
 
 use Gitlab\Api\ApiInterface;
+use Gitlab\HttpClient\Message\ResponseMediator;
 
 /**
  * Pager class for supporting pagination in Gitlab classes
@@ -33,9 +34,7 @@ class ResultPager implements ResultPagerInterface
      */
     public function fetch(ApiInterface $api, $method, array $parameters = array())
     {
-        $result = call_user_func_array(array($api, $method), $parameters);
-
-        return $result;
+        return call_user_func_array(array($api, $method), $parameters);
     }
 
     /**
@@ -44,7 +43,6 @@ class ResultPager implements ResultPagerInterface
     public function fetchAll(ApiInterface $api, $method, array $parameters = array())
     {
         $result = call_user_func_array(array($api, $method), $parameters);
-
         while ($this->hasNext()) {
             $result = array_merge($result, $this->fetchNext());
         }
@@ -105,7 +103,17 @@ class ResultPager implements ResultPagerInterface
      */
     protected function has($key)
     {
-        return !empty($this->client->getHttpClient()->getLastResponse()->getPagination()) && isset($this->client->getHttpClient()->getLastResponse()->getPagination()[$key]);
+        $lastResponse = $this->client->getResponseHistory()->getLastResponse();
+        if ($lastResponse == null) {
+            return false;
+        }
+
+        $pagination = ResponseMediator::getPagination($lastResponse);
+        if ($pagination == null) {
+            return false;
+        }
+
+        return isset($pagination[$key]);
     }
 
     /**
@@ -113,9 +121,12 @@ class ResultPager implements ResultPagerInterface
      */
     protected function get($key)
     {
-        if ($this->has($key)) {
-            $result = $this->client->getHttpClient()->get(strtr($this->client->getHttpClient()->getLastResponse()->getPagination()[$key], array($this->client->getBaseUrl() => '')))->getContent();
-            return $result;
+        if (!$this->has($key)) {
+            return [];
         }
+
+        $pagination = ResponseMediator::getPagination($this->client->getResponseHistory()->getLastResponse());
+
+        return ResponseMediator::getContent($this->client->getHttpClient()->get($pagination[$key]));
     }
 }
