@@ -2,7 +2,7 @@
 
 use Gitlab\Api\AbstractApi;
 
-class GroupsTest extends ApiTestCase
+class GroupsTest extends TestCase
 {
     /**
      * @test
@@ -21,7 +21,7 @@ class GroupsTest extends ApiTestCase
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->all(1, 10));
+        $this->assertEquals($expectedArray, $api->all(['page' => 1, 'per_page' => 10]));
     }
 
     /**
@@ -37,51 +37,11 @@ class GroupsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('groups', array('page' => 1, 'per_page' => AbstractApi::PER_PAGE))
+            ->with('groups', array())
             ->will($this->returnValue($expectedArray))
         ;
 
         $this->assertEquals($expectedArray, $api->all());
-    }
-
-    /**
-     * @test
-     */
-    public function shouldSearchGroups()
-    {
-        $expectedArray = array(
-            array('id' => 1, 'name' => 'A group'),
-            array('id' => 2, 'name' => 'Another group'),
-        );
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('groups?search=some%20group', array('page' => 1, 'per_page' => AbstractApi::PER_PAGE))
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->search('some group'));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldSearchGroupsWithPagination()
-    {
-        $expectedArray = array(
-            array('id' => 1, 'name' => 'A group'),
-            array('id' => 2, 'name' => 'Another group'),
-        );
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('groups?search=group', array('page' => 2, 'per_page' => 5))
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->search('group', 2, 5));
     }
 
     /**
@@ -111,7 +71,7 @@ class GroupsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('post')
-            ->with('groups', array('name' => 'A new group', 'path' => 'a-new-group', 'description' => null, 'visibility_level' => 0))
+            ->with('groups', array('name' => 'A new group', 'path' => 'a-new-group', 'description' => null, 'visibility' => 'private'))
             ->will($this->returnValue($expectedArray))
         ;
 
@@ -121,18 +81,18 @@ class GroupsTest extends ApiTestCase
     /**
      * @test
      */
-    public function shouldCreateGroupWithDescriptionAndVisLevel()
+    public function shouldCreateGroupWithDescriptionAndVisibility()
     {
         $expectedArray = array('id' => 1, 'name' => 'A new group', 'visibility_level' => 2);
 
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('post')
-            ->with('groups', array('name' => 'A new group', 'path' => 'a-new-group', 'description' => 'Description', 'visibility_level' => 2))
+            ->with('groups', array('name' => 'A new group', 'path' => 'a-new-group', 'description' => 'Description', 'visibility' => 'public'))
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->create('A new group', 'a-new-group', 'Description', 2));
+        $this->assertEquals($expectedArray, $api->create('A new group', 'a-new-group', 'Description', 'public'));
     }
 
     /**
