@@ -3,7 +3,7 @@
 use Gitlab\Api\AbstractApi;
 use Gitlab\Api\Projects;
 
-class ProjectsTest extends ApiTestCase
+class ProjectsTest extends TestCase
 {
     /**
      * @test
@@ -12,7 +12,7 @@ class ProjectsTest extends ApiTestCase
     {
         $expectedArray = $this->getMultipleProjectsData();
 
-        $api = $this->getMultipleProjectsRequestMock('projects/all', $expectedArray);
+        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray);
 
         $this->assertEquals($expectedArray, $api->all());
     }
@@ -24,9 +24,9 @@ class ProjectsTest extends ApiTestCase
     {
         $expectedArray = $this->getMultipleProjectsData();
 
-        $api = $this->getMultipleProjectsRequestMock('projects/all', $expectedArray, 1, 5, 'name', 'asc');
+        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray, ['page' => 1, 'per_page' => 5, 'order_by' => 'name', 'sort' => 'asc']);
 
-        $this->assertEquals($expectedArray, $api->all(1, 5, 'name'));
+        $this->assertEquals($expectedArray, $api->all(['page' => 1, 'per_page' => 5, 'order_by' => 'name', 'sort' => 'asc']));
     }
 
     /**
@@ -39,7 +39,7 @@ class ProjectsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('projects/all', array('page' => 1, 'per_page' => AbstractApi::PER_PAGE, 'order_by' => Projects::ORDER_BY, 'sort' => Projects::SORT))
+            ->with('projects')
             ->will($this->returnValue($expectedArray))
         ;
 
@@ -53,9 +53,9 @@ class ProjectsTest extends ApiTestCase
     {
         $expectedArray = $this->getMultipleProjectsData();
 
-        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray, 2, 7);
+        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray);
 
-        $this->assertEquals($expectedArray, $api->accessible(2, 7));
+        $this->assertEquals($expectedArray, $api->all());
     }
 
     /**
@@ -65,9 +65,9 @@ class ProjectsTest extends ApiTestCase
     {
         $expectedArray = $this->getMultipleProjectsData();
 
-        $api = $this->getMultipleProjectsRequestMock('projects?owned=1', $expectedArray, 3, 50);
+        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray, ['owned' => 'true']);
 
-        $this->assertEquals($expectedArray, $api->owned(3, 50));
+        $this->assertEquals($expectedArray, $api->all(['owned' => true]));
     }
 
     /**
@@ -77,14 +77,8 @@ class ProjectsTest extends ApiTestCase
     {
         $expectedArray = $this->getMultipleProjectsData();
 
-        $api = $this->getMultipleProjectsRequestMock('projects/search/a%20project', $expectedArray);
-        $this->assertEquals($expectedArray, $api->search('a project'));
-
-        $api = $this->getMultipleProjectsRequestMock('projects/search/a%2Eproject', $expectedArray);
-        $this->assertEquals($expectedArray, $api->search('a.project'));
-
-        $api = $this->getMultipleProjectsRequestMock('projects/search/a%2Fproject', $expectedArray);
-        $this->assertEquals($expectedArray, $api->search('a/project'));
+        $api = $this->getMultipleProjectsRequestMock('projects', $expectedArray, ['search' => 'a project']);
+        $this->assertEquals($expectedArray, $api->all(['search' => 'a project']));
     }
 
     /**
@@ -211,99 +205,6 @@ class ProjectsTest extends ApiTestCase
         ;
 
         $this->assertEquals($expectedBool, $api->remove(1));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGetBuilds()
-    {
-        $expectedArray = array(
-            array('id' => 1, 'status' => 'success'),
-            array('id' => 2, 'status' => 'failed')
-        );
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('projects/1/builds')
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->builds(1));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGetBuildsWithScope()
-    {
-        $expectedArray = array(
-            array('id' => 1, 'status' => 'success'),
-        );
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('projects/1/builds', array('scope' => 'success'))
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->builds(1, 'success'));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGetBuildsWithMultipleScopes()
-    {
-        $expectedArray = array(
-            array('id' => 1, 'status' => 'success'),
-            array('id' => 1, 'status' => 'failed'),
-        );
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('projects/1/builds', array('scope' => array('success', 'failed')))
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->builds(1, array('success', 'failed')));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGetBuild()
-    {
-        $expectedArray = array('id' => 2);
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('projects/1/builds/2')
-            ->will($this->returnValue($expectedArray))
-        ;
-
-        $this->assertEquals($expectedArray, $api->build(1, 2));
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGetTrace()
-    {
-        $expectedString = 'runner trace of a specific build';
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('get')
-            ->with('projects/1/builds/2/trace')
-            ->will($this->returnValue($expectedString))
-        ;
-
-        $this->assertEquals($expectedString, $api->trace(1, 2));
     }
 
     /**
@@ -637,7 +538,7 @@ class ProjectsTest extends ApiTestCase
     /**
      * @test
      */
-    public function shouldGetKeys()
+    public function shouldGetDeployKeys()
     {
         $expectedArray = array(
             array('id' => 1, 'title' => 'test-key'),
@@ -647,28 +548,28 @@ class ProjectsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('projects/1/keys')
+            ->with('projects/1/deploy_keys')
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->keys(1));
+        $this->assertEquals($expectedArray, $api->deployKeys(1));
     }
 
     /**
      * @test
      */
-    public function shouldGetKey()
+    public function shouldGetDeployKey()
     {
         $expectedArray = array('id' => 2, 'title' => 'another-key');
 
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('projects/1/keys/2')
+            ->with('projects/1/deploy_keys/2')
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->key(1, 2));
+        $this->assertEquals($expectedArray, $api->deployKey(1, 2));
     }
 
     /**
@@ -681,62 +582,45 @@ class ProjectsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('post')
-            ->with('projects/1/keys', array('title' => 'new-key', 'key' => '...'))
+            ->with('projects/1/deploy_keys', array('title' => 'new-key', 'key' => '...'))
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->addKey(1, 'new-key', '...'));
+        $this->assertEquals($expectedArray, $api->addDeployKey(1, 'new-key', '...'));
     }
 
     /**
      * @test
      */
-    public function shouldRemoveKey()
+    public function shouldDeleteDeployKey()
     {
         $expectedBool = true;
 
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('delete')
-            ->with('projects/1/keys/3')
+            ->with('projects/1/deploy_keys/3')
             ->will($this->returnValue($expectedBool))
         ;
 
-        $this->assertEquals($expectedBool, $api->removeKey(1, 3));
+        $this->assertEquals($expectedBool, $api->deleteDeployKey(1, 3));
     }
 
     /**
      * @test
      */
-    public function shoudEnableKey()
+    public function shoudEnableDeployKey()
     {
         $expectedBool = true;
 
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('post')
-            ->with('projects/1/keys/3/enable')
+            ->with('projects/1/deploy_keys/3/enable')
             ->will($this->returnValue($expectedBool))
         ;
 
-        $this->assertEquals($expectedBool, $api->enableKey(1, 3));
-    }
-
-    /**
-     * @test
-     */
-    public function shoudDisableKey()
-    {
-        $expectedBool = true;
-
-        $api = $this->getApiMock();
-        $api->expects($this->once())
-            ->method('delete')
-            ->with('projects/1/keys/3/disable')
-            ->will($this->returnValue($expectedBool))
-        ;
-
-        $this->assertEquals($expectedBool, $api->disableKey(1, 3));
+        $this->assertEquals($expectedBool, $api->enableDeployKey(1, 3));
     }
 
     /**
@@ -752,10 +636,7 @@ class ProjectsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('projects/1/events', array(
-                'page' => 1,
-                'per_page' => AbstractApi::PER_PAGE
-            ))
+            ->with('projects/1/events', array())
             ->will($this->returnValue($expectedArray))
         ;
 
@@ -782,7 +663,7 @@ class ProjectsTest extends ApiTestCase
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->events(1, 2, 15));
+        $this->assertEquals($expectedArray, $api->events(1, ['page' => 2, 'per_page' => 15]));
     }
 
     /**
@@ -1024,12 +905,12 @@ class ProjectsTest extends ApiTestCase
         $this->assertEquals($expectedBool, $api->removeVariable(1, 'ftp_password'));
     }
 
-    protected function getMultipleProjectsRequestMock($path, $expectedArray = array(), $page = 1, $per_page = 20, $order_by = 'created_at', $sort = 'asc')
+    protected function getMultipleProjectsRequestMock($path, $expectedArray = array(), $expectedParameters = array())
     {
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with($path, array('page' => $page, 'per_page' => $per_page, 'order_by' => $order_by, 'sort' => $sort))
+            ->with($path, $expectedParameters)
             ->will($this->returnValue($expectedArray))
         ;
 
@@ -1049,10 +930,7 @@ class ProjectsTest extends ApiTestCase
         $api = $this->getApiMock();
         $api->expects($this->once())
             ->method('get')
-            ->with('projects/1/deployments', array(
-                'page' => 1,
-                'per_page' => AbstractApi::PER_PAGE
-            ))
+            ->with('projects/1/deployments', array())
             ->will($this->returnValue($expectedArray))
         ;
 
@@ -1079,7 +957,7 @@ class ProjectsTest extends ApiTestCase
             ->will($this->returnValue($expectedArray))
         ;
 
-        $this->assertEquals($expectedArray, $api->deployments(1, 2, 15));
+        $this->assertEquals($expectedArray, $api->deployments(1, ['page' => 2, 'per_page' => 15]));
     }
 
     protected function getMultipleProjectsData()
