@@ -1,6 +1,6 @@
 <?php namespace Gitlab\Tests\Api;
 
-class SnippetsTest extends ApiTestCase
+class SnippetsTest extends TestCase
 {
     /**
      * @test
@@ -105,6 +105,26 @@ class SnippetsTest extends ApiTestCase
         ;
 
         $this->assertEquals($expectedBool, $api->remove(1, 3));
+    }
+
+    /**
+     * @test
+     */
+    public function shouldGetSnippetAwardEmoji()
+    {
+        $expectedArray = array(
+            array('id' => 1, 'name' => 'sparkles'),
+            array('id' => 2, 'name' => 'heart_eyes'),
+        );
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('get')
+            ->with('projects/1/snippets/2/award_emoji')
+            ->will($this->returnValue($expectedArray))
+        ;
+
+        $this->assertEquals($expectedArray, $api->awardEmoji(1, 2));
     }
 
     protected function getApiClass()
